backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
anthropic==0.39.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.19
python-dotenv==1.0.1
